Use the Kintone CLI to export data on Windows OS
